+++ driver-core/src/main/com/mongodb/connection/SocketStreamFactory.java
@@ -18,13 +18,12 @@ package com.mongodb.connection;
 
 import com.mongodb.MongoClientException;
 import com.mongodb.ServerAddress;
+import com.mongodb.UnixServerAddress;
 import com.mongodb.internal.connection.PowerOfTwoBufferPool;
 
 import javax.net.SocketFactory;
 import javax.net.ssl.SSLContext;
 import java.security.NoSuchAlgorithmException;
-
-import jnr.unixsocket.UnixSocketAddress;
 
 import static com.mongodb.assertions.Assertions.notNull;
 
@@ -65,18 +64,20 @@ public class SocketStreamFactory implements StreamFactory {
     @Override
     public Stream create(final ServerAddress serverAddress) {
         Stream stream;
-        if (serverAddress.getSocketAddress() instanceof UnixSocketAddress) {
-            stream = new SocketChannelStream(serverAddress, settings, sslSettings, bufferProvider);
-        } else if (socketFactory != null) {
-            stream = new SocketStream(serverAddress, settings, sslSettings, socketFactory, bufferProvider);
-        } else if (sslSettings.isEnabled()) {
-            stream = new SocketStream(serverAddress, settings, sslSettings, getSslContext().getSocketFactory(), bufferProvider);
-        } else if (System.getProperty("org.mongodb.useSocket", "false").equals("true")) {
-            stream = new SocketStream(serverAddress, settings, sslSettings, SocketFactory.getDefault(), bufferProvider);
+        if (serverAddress instanceof UnixServerAddress) {
+            if (sslSettings.isEnabled()) {
+                throw new MongoClientException("Socket based connections do not support ssl");
+            }
+            stream = new UnixSocketChannelStream((UnixServerAddress) serverAddress, settings, sslSettings, bufferProvider);
         } else {
-            stream = new SocketChannelStream(serverAddress, settings, sslSettings, bufferProvider);
+            if (socketFactory != null) {
+                stream = new SocketStream(serverAddress, settings, sslSettings, socketFactory, bufferProvider);
+            } else if (sslSettings.isEnabled()) {
+                stream = new SocketStream(serverAddress, settings, sslSettings, getSslContext().getSocketFactory(), bufferProvider);
+            } else {
+                stream = new SocketChannelStream(serverAddress, settings, sslSettings, bufferProvider);
+            }
         }
-
         return stream;
     }