+++ docs/reference/content/whats-new.md
@@ -7,6 +7,16 @@ title = "What's New"
   pre = "<i class='fa fa-level-up'></i>"
 +++
 
+## What's New in 3.7
+
+Key new features of the 3.7 Java driver release:
+
+### Unix domain socket support
+
+The 3.7 driver adds support for Unix domain sockets via the [`jnr.unixsocket`](http://https://github.com/jnr/jnr-unixsocket) library.
+Connecting to Unix domain sockets is done via the [`ConnectionString`]({{< apiref "com/mongodb/ConnectionString" >}}) or via
+[`UnixServerAddress`]({{<apiref "com/mongodb/UnixServerAddress.html">}}).
+
 ## What's New in 3.6
 
 Key new features of the 3.6 Java driver release: